+++ reduct/client.py
@@ -17,6 +17,31 @@ class Defaults(BaseModel):
     """settings for a new bucket"""
 
 
+class LicenseInfo(BaseModel):
+    """License information"""
+
+    licensee: str
+    """Licensee usually is the company name"""
+
+    invoice: str
+    """Invoice number"""
+
+    expiry_date: datetime
+    """Expiry date of the license in ISO 8601 format (UTC)"""
+
+    plan: str
+    """Plan name"""
+
+    device_number: int
+    """Number of devices (0 for unlimited)"""
+
+    disk_quota: int
+    """Disk quota in TB (0 for unlimited)"""
+
+    fingerprint: str
+    """License fingerprint"""
+
+
 class ServerInfo(BaseModel):
     """Server stats"""
 
@@ -37,6 +62,9 @@ class ServerInfo(BaseModel):
 
     latest_record: int
     """UNIX timestamp of the latest record in microseconds"""
+
+    license: Optional[LicenseInfo] = None
+    """license information"""
 
     defaults: Defaults
     """Default server settings"""